Move tests to test_subtensor_lift.py
